"contentMediaType" can describe binary resources
This expansion of contentMediaType is motivated by the need to
indicate media types for binary resources in hypermedia environments.

Existing usage (e.g. OpenAPI 3.0) considered unencoded binary data
to be strings, but such data violates the expectations of JSON strings.
A better approach is to purely indicate the media type and avoid
constraining the instance by JSON type as no JSON type is suitable.