Remove special case handling of annotation values
The approach of annotation keywords specifying combination rules
is not really compatible with the output format approach.  It is
also a source of uncertainty in how to use and specify annotations.

This removes the concept.  Note that "readOnly", "writeOnly", and
"deprecated" still advise applications on how to handle multiple
values.  However, the wording is purely in terms of application
handling and not in terms of aggregating the values, so no change
to the validation spec is necessary.