Fix 2019-09 hyperschema meta-schemas
The links.json schema uses `$recursiveRef` wrong. In fact,
`$recursiveRef` is not capable of expressing the relationship that was
intended. With this version, schemas in the `links` keyword are pinned
to the 2019-09 hyperschema dialect even if `$recursiveRef` is used to
extend the 2019-09 hyperschema dialect. While not exactly what was
intended, this covers all but advanced use-cases.