Clarify that plain name fragments are neither canonical or non-canonical (#1192)
* Clarify that plain name fragments are neither canonical or non-canonical
Attempt to resolve #937
Add note and cref in appendix A clarifying that we intended to define a URI phrasing which would avoid the requirement to allow for location shadowing in implementations, as this is tricky. Clarifying that plain name fragments should always be supported, and that they only can work in relation to the base URI of the Schema Resource. Otherwise there could be duplicate plain name fragments and addressing wouldn't work.