Specify the specific BSD and AFL versions this repo is licensed under.
Historically JSON Schema has declared that contributions are licensed
BSD or AFL, without specifying the version of BSD. We’ve been
asked to specify a version for work elsewhere in the ecosystem, and
OpenJS Foundation counsel recommended BSD 3-clause as the most common
choice. This commit clarifies the specific versions (3-clause and v3.0
respectively), and adds a LICENSE file to make this explicit.

See https://github.com/json-schema-org/json-schema-spec/issues/1160#issuecomment-1123583441
for the full guidance.

Closes: #1160