Track the xml2rfc dependency using a requirements.txt file
The standard practice for tracking dependencies on Python projects is to
declare the dependencies and the expected versions on a
`requirements.txt` file that is read by `pip`.

Right now, we request people to manually install the latest version of
`xml2rfc`. After this commit, we properly track `xml2rfc` on
`requirements.txt` and tweak the documentation a bit to show users how
to install it from there instead.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>