Ignore unknown keywords or treat as annotations
When we added the "SHOULD collect as annotations" behavior for
unknown keywords, we forgot to retain the MUST directive to ignore
them if the SHOULD is not followed.  Which also ensures that
no other behavior than collecting as a annotations is performed.

This also consolidates the behavioral specification in one place
as it appeared in two places before.