Disallow "#" as an $id.
In a document root schema, "$id": "#" is a no-op, and in
an embedded schema resource root, it results in the embeddded
resource having the same URI as its context resource.  Given
that the specificaiton says implementations SHOULD raise an error
for duplicate IRIs, there is no reason to allow this form.