Make targets for a virtualenv
It's common to make a virtualenv inside of your cloned repo
to ensure that you have the right enviornment for that repository,
and .venv is a common convention for that virtualenv.

This adds targets to create and clean a .venv virtual env using
the checked-in requirements.txt file.  It also adds the .venv
to the .gitignore, and updates the README.