Normative language for "$vocabulary"
This reworks the language around "$vocabulary" to capture the
original intent in normative language.  In particular, it explicitly
addresses vocabularies omitted from "$vocabulary".

The new text emphasizes the concepts of required or optional vocabularies
more than the true and false boolean values that implement them, and
adds some explanation of the utility of optional vocabularies.

It also adds a clarification around identically-named
vocabulary and non-vocabulary keywords, including that the standard
keywords MUST be considered vocabulary keywords and subject to
"$vocabulary" control even if the implementation does not support
custom vocabularies at all.