Reorganize plain name fragment requirements
In response to PR review feedback, remove the editorial explanation
and instead consolidate the fragment syntax in the section on
fragments.  The context and reason for the syntax is more clear
there with the existing text, and one additional line further
clarifies it by not just seeming to reference a random XML
syntax rule for no clear reason.