Remove erroneous use of "default base IRI"
This should just say "base IRI", as a "default base IRI" is a
specific kind of base IRI that does not make sense in the context
of this paragraph.

The term is used correctly several paragraphs down where there is
a RECOMMENDATION that if a default base IRI is defined (which is
not in any way required) that it be documented.