Clarify how to use "contentSchema"
We know that many schema users find base URIs/IRIs confusing, and
that the behavior of implementations in the absence of "$schema"
is inconsistent.  Without delving into those topics (which are
addressed in core, or in RFC 3986), this clarification shows how
to use "contentSchema" as a normal subschema from an annotation.

It also explains the circumstances under which the extracted
annotation value is safe to use.  The language attempts to balance
the likelihood that readers will not be familiar enough with
the restrictions to note this on their own with the desire to
minimize re-stating schema processing rules documented in the core
specification.  This should be enough to encourage readers to
check the relevant core specification sections.