Only allow non-zero index adjustmetns
The current draft allows for an index adjustment of "0" which
creates ambiguity because "1/foo", "1+0/foo", and "1-0/foo"
are now three different ways to write a pointer with the exact
same effect.

This complicates round-trips between textual and functional
representations, the original text would need to be preserved
in order to re-constitute it correctly.  There is no need for
this added complication.